Implement Copy-Paste for all Textfields

+++ OpenBCI_GUI/W_Playback.pde
@@ -90,6 +90,13 @@ class W_playback extends Widget {
     }
 
     public void refreshPlaybackList() {
+
+        File f = new File(userPlaybackHistoryFile);
+        if (!f.exists()) {
+            println("OpenBCI_GUI::RefreshPlaybackList: Playback history file not found.");
+            return;
+        }
+
         try {
             playbackMenuList.items.clear();
             loadPlaybackHistoryJSON = loadJSONObject(userPlaybackHistoryFile);
@@ -251,20 +258,24 @@ boolean playbackFileSelected (String longName, String shortName) {
     outputSuccess("You have selected \""
     + shortName + "\" for playback.");
 
-    try {
-        savePlaybackHistoryJSON = loadJSONObject(userPlaybackHistoryFile);
-        JSONArray recentFilesArray = savePlaybackHistoryJSON.getJSONArray("playbackFileHistory");
-        playbackHistoryFileExists = true;
-    } catch (NullPointerException e) {
-        //println("Playback history JSON file does not exist. Load first file to make it.");
+    File f = new File(userPlaybackHistoryFile);
+    if (!f.exists()) {
+        println("OpenBCI_GUI::playbackFileSelected: Playback history file not found.");
         playbackHistoryFileExists = false;
-    } catch (RuntimeException e) {
-        outputError("Found an error in UserPlaybackHistory.json. Deleting this file. Please, Restart the GUI.");
-        File file = new File(userPlaybackHistoryFile);
-        if (!file.isDirectory()) {
-            file.delete();
+    } else {
+        try {
+            savePlaybackHistoryJSON = loadJSONObject(userPlaybackHistoryFile);
+            JSONArray recentFilesArray = savePlaybackHistoryJSON.getJSONArray("playbackFileHistory");
+            playbackHistoryFileExists = true;
+        } catch (RuntimeException e) {
+            outputError("Found an error in UserPlaybackHistory.json. Deleting this file. Please, Restart the GUI.");
+            File file = new File(userPlaybackHistoryFile);
+            if (!file.isDirectory()) {
+                file.delete();
+            }
         }
     }
+    
     //add playback file that was processed to the JSON history
     savePlaybackFileToHistory(longName);
     return true;
